Move to autotools, add standard CI automation scripts
Move packaging out of an additional checkout in a subdir to a
Makefile, with associated tooling (autoconf, etc) to set the
version in one place and manage it everywhere.

Fix some linter errors in the javascript so it builds/lints
cleanly.

Add standard CI automation scripts for linting on patch/merge and
building artifacts for export.

Prep everything for a merge with the node dashboard

Change-Id: Id85914fae711d5b6a96c58ef3c4c3a1ac902e870
Signed-off-by: Ryan Barry <rbarry@redhat.com>