Move node-dashboard hosted engine components to react
Bring the hosted-engine setup page to parity with the angular
version -- dynamically fill the input, use patternfly styling
for warnings/info, supply error helper text if the input is
invalid, and flag the input field as bad.

Do hosted-engine-setup in a helper to keep the state of the views
clean.

Fix up the packaging to build things properly.

Move all the React stuff to es6 classes

Don't use webpack for cockpit, because it doesn't attach to dbus
and other services unless we require it on the page, it seems

TODO:

Add a basic dashboard page

(Eventually) move the hosted-engine communication parts to Redux

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1298160

Change-Id: Ie1195d7319998dcf7e11196aa107484bc42e5e25
Signed-off-by: Ryan Barry <rbarry@redhat.com>