Add some description to the Hosted Engine page, fix up index
Read the cockpit source. Full functionality is available if jquery
is imported. Remove the requirejs bits, and rely on the jquery
import to provide full functionality.

Add an informational message to the hosted engine page.

Unmount the timer in the management page so it won't keep trying
to update state if a different tab is selected

Change-Id: I7a2393bad97655413358f715efabca036c2a1976
Signed-off-by: Ryan Barry <rbarry@redhat.com>