Add content to the dashboard
Add running VMs back.

Pull in all of patternfly so we can use full styling in the
dashboard, even though this makes it slightly larger.

Discard the merge with the vdsm codebase, since that will no
longer be maintained (we'll keep it in a tab as long as it works
or until 4.1)

Add a recursive mapper for NetworkManager. Keep track of seen
keys, since the objects are recursive trees, and NetworkManager
creates new objects every time a connection is changed, so we don't
need to worry about collisions.

Coming: put the networkmanager model on the page, SSH host key,
uglify/minify/dedupe

Change-Id: I25c75fb424229c16c2a7a3cf0db223166fefcf32
Signed-off-by: Ryan Barry <rbarry@redhat.com>