Use stable Cockpit CSS identifiers
As part of stabilizing Cockpit and allowing out of tree components
to be built for it, we're cleaning up the CSS.

In particular adding a -ct namespace to all CSS identifiers. This
follows the scheme setup by Patternfly with their -pf namespace.

I've left the old classes in place so that cockpit-ovirt will
be compatible with older and future Cockpit releases.

Change-Id: I7dfdd6ec7ebde0cc3195c0e35fafe8c271332505
Signed-off-by: Stef Walter <stefw@redhat.com>