Define the CSS class content-header-extra ourselves
As part of stabilizing Cockpit and allowing out of tree components
to be built for it, we're cleaning up the CSS.

In this case we've removed the content-header-extra definition
which won't be used in Cockpit upstream across components.

I've renamed and added the definition here, so that cockpit-ovirt
will be compatible with older and future Cockpit releases.

Change-Id: Ie0fc8dd113f9159a68f9374e54dba8510ee6bb7e
Signed-off-by: Stef Walter <stefw@redhat.com>