dashboard: Add hosted-engine status
Show a spinner while we wait to see whether hosted engine has
been deployed. If it has, wait while we check the status.

If the engine isn't running, show a warning.

If it is running, show what host it's running on, then pull
metrics from all the hosts and show them.

Put the hosts into a grid, with two entries per row.

Buttons for maintenance modes coming...

Change-Id: I71947014b91974940ef102758e27b7e1cc9dd1e2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1318415
Signed-off-by: Ryan Barry <rbarry@redhat.com>