dashboard: UX improvements
Use a table for the links on the dashboard page.

Add some titles and headers to the hosted engine status page

Don't use patternfly listview, so we avoid onhover events. But add
a couple of extra CSS classes to keep vertical alignment and
margins.

Move the list of running VMs to another column in a grid

Change-Id: Ie9ba0120d1c172209f408ac98af781757675637d
Signed-off-by: Ryan Barry <rbarry@redhat.com>