packaging: decouple testing and building
Avoid to run npm test on build phase.
Tests are meant to be executed if make check is called.
The call to make should just build.

As side effect of the above change, need to move some
make targets to -local in order to ensure that all the
dependencies are correctly resolved.

Change-Id: I6978c57615f6b5ca4277d064c7b2a39deb48f1ba
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>