dashboard: disable node calls if nodectl isn't present
In case the dashboard is installed on a regular system, check for
nodectl before rendering the table. If it doesn't exist, don't
add it to the DOM (none of the children will be called, nodectl
will never be invoked again).

Annoyingly, React requires a containing element for everything, and
we're inside of a table, so create a trivial function that returns
the previous elements (inside the tbody) as an array, which React
can deal with since we can't use a <div> inside of <tbody>

Change-Id: Ifdc5c6b91a82bb30af30e3aae9ed9daf9dbdc03f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1380641
Signed-off-by: Ryan Barry <rbarry@redhat.com>