gdeploy: add gdeploy wizard for cockpit-ovirt
Integrating gluster deployment with hosted engine deployment flow.
User will be given an option to start HE deployment with gluster.
It will open up a wizard for gdeploy configurations and user has to
give gluster node details and other input required for gluster
deployment. It will create a gdeploy config file and show it to the
user. Once user reviews and confirms the gdeploy config file, then
start the deployment using gdeploy. After successful completion of
the gluster deployment, user will be given an option to continue
with HE deployment.

Change-Id: I5e89e753d169093a9a3e39c89a867884f895c39c
Bug-Url: https://bugzilla.redhat.com/1325864
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>