gdeploy: disable gdeploy wizard if gdeploy is missing.
gdeploy is required for Gluster deployment. So gluster
deployment option should be disabled when gdeploy command is
not available.

Existence of gdeploy command is checked by running
'gdeploy --version'. If this command succeeds then it
means that gdeploy is installed in the system. If it
fails, then we will disable the option.

Change-Id: I7bf4ac5d37d17d6a0c3e841a96c53d6b6146349c
Bug-Url: https://bugzilla.redhat.com/1415189
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>