gdeploy: remove pool metadata size from gdeploy config.
gdeploy automatically calculates the thinpool metadata
size when it is not specified in the gdeploy config. Also
it adjusts the lv size when it auto calculates the thinpool
meta data size. So it is better to to remove the field
'poolmetadatasize' from gdeploy configuration.

Change-Id: Iae373e006ae5d0fb36cadd312f8ccc89e679c5c8
Bug-Url: https://bugzilla.redhat.com/1426494
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
(cherry picked from commit 2161ff6017a71ed8a219c89dfccb621834334787)