gdeploy: change volume options
with gluster bug bz#1418901, most of the gluster volume options
required for hc is set by virt profile. So we can remove the
redundant options.

Note: by default shard-block-size is 4MB when features.shard is
enabled and it is recommanded to keep shard size as 4MB for HC.
So we can remove the shard size option from the volume option
list.

Change-Id: I0c84c63de581fdcdb7fdaf71bcc01e789fd7509c
Bug-Url: https://bugzilla.redhat.com/1426527
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>