he: unset confirm_otopi after passing input
Since we need to send a bare otopi verb if the GPG_KEY needs
confirmation, we set this value if we find CONFIRM in the line.

This should be un-set after the input is passed back in, but was
not previously.

Change-Id: Id1ea972f818e6135ad855a18ffffa83dc1a45e95
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1438604
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit 895e5e7b28a546169900d17c7be00c90c63599bb)