gdeploy: Edit hc-conf file to start chronyd instead of ntpd
ntp was earlier pulled in as dependency of vdsm whereas, now from ovirt 4.1.2
it drops ntp in favour of chronyd. In gdeploy conf file there is a section to
enable and restart ntpd and this has to be changed to enable and start chronyd.

Bug Url: https://bugzilla.redhat.com/show_bug.cgi?id=1450152
Change-Id: I72b468a46adcb55b228e235a8822b4b051748af4
Signed-off-by: Devyani Kota <devyanikota@gmail.com>