wizard: Update look/feel of MultiRowTextBox component
This patch updates the look and feel of the MultiRowTextBox to be more
like that of the similar component in ovirt-engine. It also adds
additional behavior to disable the add button when there is no input in
the last row.

Change-Id: I0b667aa6a62e52945eb1b58c8793aaddacfbc33a
Signed-off-by: Phillip Bailey <phbailey@redhat.com>
(cherry picked from commit 373ba39d395b53431f5bc72f04ef305d738d4101)