wizard: Prevent ansible errors
This patch removes the AnsibleUtil class and uses the bare cockpit.spawn
call in HeSetupWizard, instead. The call to the setSystemData method is
moved to the done callback, preventing possible truncating of data. The
getTaskData method is exported from HostedEngineSetupUtil as it used
across more than one file.

Additionally, tasks are updated to ignore errors to prevent the return
JSON from being corrupted with error messages and warnings.

Bug-Url: https://bugzilla.redhat.com/1509775
Change-Id: I38008e6e81a78e85b96845bab903b7f19752924d
Signed-off-by: Phillip Bailey <phbailey@redhat.com>