wizard: Handle CPU types not supported by setup
This patch adds functionality to handle situations where the detected
CPU model is not in the supported list in ovirt-hosted-engine-setup. If
the detected model is in the known model list (currently hard-coded and
based on [1]) then the most compatible model is selected as the default.
If the detected model is not in the known model list, then the oldest
model is selected as the default. In both cases, a warning icon with
an appropriate tooltip message is displayed.

[1] https://github.com/lago-project/lago-ost-plugin/blob
/master/ovirtlago/data/ovirt_cpu_map.yaml

Change-Id: Ia3fcb7826b068928617d484900b20b58c125528e
Bug-Url: https://bugzilla.redhat.com/1516113
Signed-off-by: Phillip Bailey <phbailey@redhat.com>
(cherry picked from commit 2ba2cf89d2fb2362c0eae8bf1ad18b980754d0dc)