wizard: Prevent password validation issue
Currently, if a user clears a previously entered password after having
confirmed it in the confirmation field, the confirmation field will be
hidden. When the user attempts to progress to the next step, validation
will fail, but the field-specific  error message (which is shown on the
confirmation field) will not be visible.

This patch clears the confirmation entry when the password entry is
cleared, preventing the issue from occurring.

Change-Id: I8e7a0703def9e828e53437bc26b8d3aff33c3187
Signed-off-by: Phillip Bailey <phbailey@redhat.com>