wizard: Correct answer file type/value for firewallManager
This patch changes the return type for the firewallManager answer file
key from boolean to either string or none depending on the return value
and sets the string value to "iptables" if iptables is to be configured.

It also modifies the way that this value and boolean values are
represented in the preview step.

Change-Id: Iee33b7964e16f3a85c5fbe5f1a1cb80111bcd425
Bug-Url: https://bugzilla.redhat.com/1522641
Signed-off-by: Phillip Bailey <phbailey@redhat.com>
(cherry picked from commit 36271e338d4b366c59de1d7b79f1b14fd0382beb)