wizard: Don't let the wizard start if libvirt is down
Naive polling of `virsh -r capabilities` meant that libvirt being
down only showed AMD CPUs, even if the CPU was actually Intel.

However, since we can't actually do anything with the wizard if
libvirt is down, it's not meaningful to run it. Block the entire
thing if libvirt is down for some reason.

We should also have ansible try to start this as part of the
playbook, so facts can be gathered reliably.

Change-Id: Icfca20a2d554134161659e0d690aa5c71dfe64ce
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1528818
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1541273
Signed-off-by: Ryan Barry <rbarry@redhat.com>