wizard: Cleanup handling of VM memory defaults
This patch sets the initial value of the "Memory Size" field to the
maximum amount of available memory if it is less than the minimum
recommended amount; otherwise, it sets it to the minimum recommended
amount.

It also uses the constant for the minimum recommended amount wherever
possible and refactors the code used to set the defaults to only call
the defaultProvider methods once per execution.

Change-Id: I143cf0c94ce8979ff8ff306818d970ef6888017d
Signed-off-by: Phillip Bailey <phbailey@redhat.com>