wizard: Cleanup handling of VM memory defaults
This patch sets the initial value of the "Memory Size" field to the
maximum amount of available memory if it is less than the minimum
recommended amount; otherwise, it sets it to the minimum recommended
amount.

It also uses the constant for the minimum recommended amount wherever
possible and refactors the code used to set the defaults to only call
the defaultProvider methods once per execution.

Change-Id: I5a0876835078ef7c228c072e9ce150738598448a
Signed-off-by: Phillip Bailey <phbailey@redhat.com>