wizard: Consolidate network and VM steps
This patch moves the bridge interface, bridge name, and gateway address
fields from the network step to the VM step and removes the network step
from the wizard. It also moves execution of the playbook associated with
the network step (create_target_vm.yml) to occur after execution of
create_storage_domain.yml.

Change-Id: I649ff61a8fbeb2b021b5223736db05aeccc10fbc
Bug-Url: https://bugzilla.redhat.com/1539560
Signed-off-by: Phillip Bailey <phbailey@redhat.com>