wizard: Add prompt when gluster config is found
This patch adds functionality to display a popup dialog when gluster
configuration files are found from a previous execution of the gluster
wizard. The popup gives the user the option to use the existing
configuration or go through the gluster wizard again.

It also makes gluster the selected domain type when the user selects
the hyperconverged option from the start page.

Change-Id: I7e739b74e1aa4f6ab6360652009d8e40058b2c20
Signed-off-by: Phillip Bailey <phbailey@redhat.com>