wizard: Clear main error message in storage step
This patch clears the main error message in the storage step when the
user makes a change. The error shouldn't be displayed again unless the
user attempts to progress to the next step with invalid and/or missing
inputs that are required.

Change-Id: I9d328cbaae5e02ade85ad6247d8bf197f4e1cc7a
Signed-off-by: Phillip Bailey <phbailey@redhat.com>
(cherry picked from commit 7d0d9ec461f9106b4ae732f9e18d41d045d4110d)