wizard: Use maskable password component in VM & engine steps
This patch replaces the admin and root password fields in the VM and
engine steps with a password component that allows the password to be
unmasked. The associated confirm password fields, accompanying methods,
and fields in the model are also removed, as they are no longer
necessary.

Change-Id: Idc356874fc5725721784bc28ee16adb7695e017f
Signed-off-by: Phillip Bailey <phbailey@redhat.com>
(cherry picked from commit d3cba72e543436168f572418a3bb106d8a2c2d90)