wizard: Confirm close request in HE wizard
Currently, exiting the HE wizard using the "x", "Cancel", or "Close"
exits the wizard with no warning and all inputs are lost. This patch
adds a confirmation prompt that is displayed when the user clicks any of
these buttons. Selecting "Yes" at the prompt will exit the wizard, as
expected. Selecting "No" will return the user to the wizard with all
inputs intact.

Change-Id: I71a068f13a353ad94fecc3129cbc4dc22a1314fb
Bug-Url: https://bugzilla.redhat.com/1558082
Signed-off-by: Phillip Bailey <phbailey@redhat.com>