wizard: Align "Configure LV Cache" to reflect UI heirarchy
Previously, the UI alignment incorrectly implied that the toggle's
action was independent of the host selected in the selection box
because it was aligned at the same level as the "Brick Configuration"
section title.

Now, the toggle is aligned with the bricks table to show that it, and
the settings it toggles, are host-specific.

Change-Id: Id9cec8220215d297fe206f9b5562649e310ba121
Bug-Url: https://bugzilla.redhat.com/1560405
Signed-off-by: Rohan CJ <rojoseph@redhat.com>
(cherry picked from commit eadca7db527b63d4ea25a6020d71ef604529a61f)