Correcting the mount point generated in Brick tab
There was a condition checking for brick directories in the volume
tab which was leading to wrong mount point in the bricks in certain use-cases.
Only brick directories in format of "/gluster_bricks/example/example
had correct mount points i.e. "/gluster_bricks/example".
Also added validation for ensuring the brick directory in a certain format.

Change-Id: Ic03c266ca1f725e9f4bc30f31be577ee2e53181a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1597255
Signed-off-by: parthdhanjal <dparth@redhat.com>