Checkbox on FQDN Tab in case hostname is same as FQDN
Adding a checkbox on FQDN Tab in case the hostname and fqdn for
host 2 and host 3 are same.
The UI changes screenshots - https://imgur.com/a/G1PQKkK
On checking the optiion both the FQDN fields are disabled and
data is auto-populated in the fields. The same changes are also reflected
in the glusterModel. And upon unchecking, the fields are enabled and emptied,
also the changes are reverted from the glusterModel.

Change-Id: I0bce458f30ccd15e9483a5728f29ac7a27d43a68
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1630689
Signed-off-by: parthdhanjal <dparth@redhat.com>