Autopopulating the second and third host with the same device name for LVM Cache configuration
Autopopulating the device name under the LV Cache column for 2nd
and 3rd host with the device mentioned under the 1st host LV Cache.
Making any changes to the 1st host will be reflected to the 2nd
and the 3rd host. But any changes made to 2nd and 3rd host will be
visible only on the same.

The UI changes can be seen here - https://imgur.com/a/7os2ydM

Change-Id: I844eb5a8a870b494df957ca5110afb900d6b99cf
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1638679
Signed-off-by: parthdhanjal <dparth@redhat.com>