Configure PermitRootLogin later
Root login to the engine VM over ssh is needed
at engine-setup time because now it's executed
via an ansible role.
The user can still want to diable it so
let's temporary enable it for setup execution
and configure it via PermitRootLogin
as latest action on the engine VM before
restarting it from the shared storage.

Requires: https://github.com/oVirt/ovirt-ansible-hosted-engine-setup/pull/251

Change-Id: Ic50616397a035fd04be332dbcf78711bd3ff11dd
Bug-Url: https://bugzilla.redhat.com/1752113
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>