VGName and Thinpool name should be less than 55 characters
The problem occurs when the lv name, vg name, thin pool name is
greater than 55 characters. python-blivet throws an error here.
And in case lv cache is longer than 120 characters it throws an
error during deployment

sdd                                             8:48   0   1.8T  0 disk
sde                                             8:64   0   1.8T  0 disk
└─36000304801397f830334125618b1f1bc47213fb    253:0    0   1.8T  0 mpath

In this example above the length of mpath is greater than 39 characters
So when we add the suffix gluster_vg_<mpath>, this will be greater than
55 characters, which will cause an error in blivet
As we set lv name, vg name and thinpool name by taking the device
name as a placeholder we just need to come up with a unique character
combination from the long WWID/UUID name. We can use the
last 24 characters which are unique placeholder for these names.

After making the required changes, the output will be
sdd                                                 8:48   0   1.8T  0 disk
sde                                                 8:64   0   1.8T  0 disk
└─36000304801397f830334125618b1f1bc47213fb        253:0    0   1.8T  0 mpath
  ├─gluster_vg_30334125618b1f1bc47213fb_lv_engine 253:5    0   100G  0 lvm

Change-Id: I97c967e91c671642a345e4c4610a1c546a680933
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1884223
Signed-off-by: parthdhanjal <dparth@redhat.com>